HEEDLS-439 Split CustomPromptsService and CustomPromptsDataService into separate services for different kinds of prompts
