+++ DigitalLearningSolutions.Data.Tests/Services/CourseServiceTests.cs
@@ -5,6 +5,7 @@
     using DigitalLearningSolutions.Data.DataServices;
     using DigitalLearningSolutions.Data.Models.Courses;
     using DigitalLearningSolutions.Data.Services;
+    using DigitalLearningSolutions.Data.Services.CustomPromptsService;
     using FakeItEasy;
     using FluentAssertions;
     using NUnit.Framework;
