+++ DigitalLearningSolutions.Data/Services/CourseService.cs
@@ -4,6 +4,7 @@
     using System.Linq;
     using DigitalLearningSolutions.Data.DataServices;
     using DigitalLearningSolutions.Data.Models.Courses;
+    using DigitalLearningSolutions.Data.Services.CustomPromptsService;
 
     public interface ICourseService
     {
